+++ src/App.css
@@ -96,7 +96,6 @@ a {
   text-decoration: none;
   display: flex;
   align-items: center;
-  height: 100%;
   letter-spacing: 3px;
 }
 span:hover {
@@ -108,7 +107,7 @@ span:hover {
 .banner-main {
   background: #e2b659;
   position: relative;
-  height: 600px;
+  height: 100vh;
 }
 .banner-main > * {
   position: absolute;
@@ -134,6 +133,7 @@ h2 {
 .section {
   text-align: center;
   padding: 6em 0;
+  height: 100vh;
 }
 
 .section > h3 {
@@ -228,6 +228,10 @@ ul {
   color: #fff;
   text-align: center;
   background: #000;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
 }
 
 a.mailTo {
@@ -235,7 +239,7 @@ a.mailTo {
   border: 2px solid white;
   width: max-content;
   padding: 1rem;
-  margin: auto;
+  margin-top: 5%;
   min-width: 200px;
 }
 .links {
@@ -268,7 +272,7 @@ a.mailTo {
     padding: 7px;
   }
   .banner-main {
-    height: 400px;
+    height: 100vh;
   }
   .banner.banner-waves {
     height: 50px;
