+++ src/App.js
@@ -1,13 +1,26 @@
 import React from 'react';
 import { Route, BrowserRouter,Switch } from 'react-router-dom'
 import Main from "./Main";
+import About from "./About";
+import Contact from "./Contact";
+import "./App.css";
+
+// google analytics
 import "./ga";
 
 function App() {
   return (<>
+    <a
+      className="download-btn"
+      href="https://drive.google.com/file/d/1bfcp9no13ONLMbQowGJA6iaX15uoxM4x/view?usp=sharing"
+      download="resume_amanjotSingh"
+      ><div>CV</div>
+    </a>
     <BrowserRouter>
       <Switch>
-        <Route path="/" component={Main} />
+        <Route exact path="/" component={Main} />
+        <Route exact path="/about" component={About} />
+        <Route exact path="/contact" component={Contact} />
       </Switch>
     </BrowserRouter>
   </>
