+++ src/index.css
@@ -1,11 +1,3 @@
-body {
-  margin: 0;
-  font-family: "Roboto", sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  overflow-x: hidden;
-}
-
 @font-face {
   font-family: "Roboto";
   font-style: normal;
@@ -13,15 +5,21 @@ body {
   font-display: swap;
   src: url("https://fonts.googleapis.com/css2?family=Roboto:wght@300;400;500&display=swap");
 }
-
 * {
   font-family: sans-serif; /* sans-serif will be used as temporary font until Roboto is available */
-}
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
-    monospace;
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
+  font-size: 18px;
 }
 
+body {
+  margin: 0;
+  font-family: "Roboto", sans-serif;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  overflow-x: hidden;
+}
 .App {
   display: grid;
 }
